tt: rename "engines" package
Rename "engines" package to "templates", because this
name better reflects the essence of the package.
